+++ ergo/ChangeLog.md
@@ -6,6 +6,10 @@
   `ERGO_LOG_FILE` environment variable to override this.
 * Improve error logging to accurately track pending errors.
 * Add `-d`/`--doc` flags for getting documentation for the final value.
+* Add `-p`/`--page` flags for paging the final output.
+* Improve `ergo` use in scripts to not always apply functions when they are
+  loaded. Instead, the final value at the top level will, if a function, be
+  applied (with no arguments) to refine the final value.
 
 ## 1.0.0-beta.4  -- 2020-11-17
 * Add the `!` operator, which may precede any expression or argument to evaluate